Feat: fix sonar code smells

+++ src/main/java/com/technical/kardex/domain/handlers/ProductManagerHandler.java
@@ -34,7 +34,8 @@ public class ProductManagerHandler implements ProductManager {
     @Override
     public void addStockToProduct(StockUpdateRequest stockUpdateRequest) throws ProductNotFoundException, InvalidStockException {
         Optional<Product> optionalProduct = productRepository.findById(stockUpdateRequest.getProductId());
-        optionalProduct.orElseThrow(() -> new ProductNotFoundException());
+        if(!optionalProduct.isPresent())
+            throw new ProductNotFoundException();
         Product product = optionalProduct.get();
         product.setStock(product.getStock() + stockUpdateRequest.getStock());
         validateStock(product);
@@ -44,7 +45,8 @@ public class ProductManagerHandler implements ProductManager {
     @Override
     public void updateTotalStockProduct(StockUpdateRequest stockUpdateRequest) throws ProductNotFoundException, InvalidStockException {
         Optional<Product> optionalProduct = productRepository.findById(stockUpdateRequest.getProductId());
-        optionalProduct.orElseThrow(() -> new ProductNotFoundException());
+        if(!optionalProduct.isPresent())
+            throw new ProductNotFoundException();
         Product product = optionalProduct.get();
         product.setStock(stockUpdateRequest.getStock());
         validateStock(product);
@@ -63,7 +65,8 @@ public class ProductManagerHandler implements ProductManager {
     @Override
     public Product getProductById(long productId) throws ProductNotFoundException {
         Optional<Product> optionalProduct = productRepository.findById(productId);
-        optionalProduct.orElseThrow(() -> new ProductNotFoundException());
+        if(!optionalProduct.isPresent())
+            throw new ProductNotFoundException();
         return optionalProduct.get();
     }
 
@@ -80,7 +83,8 @@ public class ProductManagerHandler implements ProductManager {
     @Override
     public void updateProduct(ProductInformation product) throws InvalidStockException, ProductNotFoundException {
         Optional<Product> optionalProduct = productRepository.findById(product.getId());
-        optionalProduct.orElseThrow(() -> new ProductNotFoundException());
+        if(!optionalProduct.isPresent())
+            throw new ProductNotFoundException();
         validateStock(product.toEntity());
         productRepository.save(product.toEntity());
     }
@@ -88,7 +92,8 @@ public class ProductManagerHandler implements ProductManager {
     @Override
     public void deleteProduct(long productId) throws ProductNotFoundException {
         Optional<Product> optionalProduct = productRepository.findById(productId);
-        optionalProduct.orElseThrow(() -> new ProductNotFoundException());
+        if(!optionalProduct.isPresent())
+            throw new ProductNotFoundException();
         productRepository.deleteById(productId);
     }
